+++ slo-workload/src/Defaults.php
@@ -4,27 +4,25 @@ namespace YdbPlatform\Ydb\Slo;
 
 class Defaults
 {
-    const TABLE_NAME = 'slo-php';
     const TABLE_MIN_PARTITION_COUNT = 6;
     const TABLE_MAX_PARTITION_COUNT = 1000;
     const TABLE_PARTITION_SIZE = 1;
 
-    const GENERATOR_DATA_COUNT = 1000;
+    const PREFILL_COUNT = 1000;
+
+    const DURATION = 600; // seconds
 
     const READ_FORKS = 3;
     const WRITE_FORKS = 1;
 
     const READ_RPS = 1000;
-    const READ_TIMEOUT = 70; // milliseconds
-    const READ_TIME = 140; // seconds
+    const READ_TIMEOUT = 10000; // milliseconds
 
     const WRITE_RPS = 100;
-    const WRITE_TIMEOUT = 20000; // milliseconds
-    const WRITE_TIME = 140; // seconds
+    const WRITE_TIMEOUT = 10000; // milliseconds
 
-    const SHUTDOWN_TIME = 30;
-    const PROMETHEUS_PUSH_GATEWAY = 'http://prometheus-pushgateway:9091';
-    const PROMETHEUS_PUSH_PERIOD = 250; // milliseconds
+    const SHUTDOWN_TIME = 30; // seconds
+    const REPORT_PERIOD = 1000; // milliseconds
 
     const WRITE_QUERY = 'DECLARE $id AS Uint64;
 DECLARE $payload_str AS Utf8;